requirements.txt
django
djangorestframework
psycopg2
django-markdownx
python-dotenv
dj-database-url
whitenoise
libsass
django-compressor
django-sass-processor
